Refactor code by separating DB-related logic
